fix: lvmd: use 0 as free bytes of thin pool if usage exceeds total capacity
In the current implementation, the free capacity of a thin pool is
calcualted by subtracting its usage from the overprovisioned total
capacity.  Since both usage and capacity are represented as unsigned
integers, the result wraps around if the usage somehow exceeds the
capacity.

This comit addresses the issue by comparing the two values before
performing the subtraction.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>